ui/util: block signal feedback loop
After typing r"\N" in the editor, it immediately replaced it to "\n" and
placed the cursor at the beginning, which is a terrible UX. The cause
for it was widget data change triggered a model data change, which was
propagated back to the widget. Now the changes are not propagated back.