ui/mpv: do not use --end
Setting mpv's --end has a history of breaking in most unusual ways.
For example, after updating to mpv 0.33.1, audio stops working upon
reaching --end, until the user performs a manual seek.

In this commit we replace the existing mechanism with a manual
implementation of --end, that might worsen precision on laggy systems,
but in general should be more stable.